requirements.txt
faiss-cpu
sentence-transformers
pypdf
